Move SDK into main hotswapper; move udev into bootstrap
More permission related shuffling. We need root for udev (to copy to /etc)
but better to install the sdk under the user.

Added a $SETUP_FILE to pass bootstrap status, and remove that file
from version control. I did that same thing with the .config in
the auto-patcher, this allows you to separate update and local
status from being pushed to remote versions.