Allow hotswapper to be called safely anywhere from filesystem
I want to remove BUILD from version control. That could be the condition that sets
bootstrap flag. It would also hardcode in a check to ensure users are not downloading
all of AOSP to a random location without realizing it. AND it allows us to have people's
personal setups not go into version control, which is much better.