Change (again) prop files
This now checks out the proper branch, allowing people to build older
versions of Android and still get the proper proprietary files.

This comes with a mjor drawback in that there will be sync errors for
those that look. They are non-fatal, and actually do not even matter.
But still, unhandled errors is no good. This needs to be sorted.

I need to find a document I can reference from within the script that
will match devices to their necessary prop files. That or remove any
non-existant prop repos. Which could be very expensive. Hmmmmm.