Only pass branch to maintained proprietary repos
This allows us to be specific when building for OEMs with maintained
prop repos. For those that are one off, I am assuming (and willing to
hear arguments) that if they are needed, they will only work for branches
of device trees they are meant for.

As in, if you are building 4.0 and need geeksphone, I am assuming that
the latest branch will cover you. These are so rarely used, it will
probably never come up. And if it does, I will spend more time handling
it then.