Add GIT_NAME and GIT_EMAIL to BUILD andorderBuilds
This is to satisfy the AOSP/git requirement for name and email. By default
the user is set as anonymous. I also added the BUILD file to .gitignore, so
that if they do choose to set their identity in BUILD, it won't
accidentally get pushed to version control. But I need BUILD to be in the
initial pull, hopefully that is not a conflicting set up.

Only way to know is to try.