Set GIT_NAME and GIT_EMAIL globally and unset after builds
We need to set these variables. I default to anonymous, but
still, even if set, I don't want to permanently set public facing variables
for users deep in the shell script. This allows us to get around it without
permanently setting any envvironmental variables.

Also pass "y" to orderbuilds, which is done to accept colors in the terminal.
A very weird insistence of interactivity by AOSP.